+++ src/Sylius/Bundle/CoreBundle/Kernel/Kernel.php
@@ -40,7 +40,6 @@ abstract class Kernel extends BaseKernel
     public function registerBundles()
     {
         $bundles = [
-            new \Sylius\Bundle\TranslationBundle\SyliusTranslationBundle(),
             new \Sylius\Bundle\InstallerBundle\SyliusInstallerBundle(),
             new \Sylius\Bundle\OrderBundle\SyliusOrderBundle(),
             new \Sylius\Bundle\MoneyBundle\SyliusMoneyBundle(),