+++ src/Sylius/Bundle/ResourceBundle/DependencyInjection/Driver/AbstractDriver.php
@@ -12,9 +12,9 @@
 namespace Sylius\Bundle\ResourceBundle\DependencyInjection\Driver;
 
 use Sylius\Component\Resource\Factory\Factory;
+use Sylius\Component\Resource\Factory\TranslatableFactoryInterface;
 use Sylius\Component\Resource\Metadata\Metadata;
 use Sylius\Component\Resource\Metadata\MetadataInterface;
-use Sylius\Component\Resource\Factory\TranslatableFactoryInterface;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Symfony\Component\DependencyInjection\Definition;
 use Symfony\Component\DependencyInjection\Parameter;
@@ -116,15 +116,12 @@ abstract class AbstractDriver implements DriverInterface
      */
     protected function addFactory(ContainerBuilder $container, MetadataInterface $metadata)
     {
-        $translatableFactoryInterface = TranslatableFactoryInterface::class;
-
         $factoryClass = $metadata->getClass('factory');
         $modelClass = $metadata->getClass('model');
 
-        $reflection = new \ReflectionClass($factoryClass);
         $definition = new Definition($factoryClass);
 
-        if (interface_exists($translatableFactoryInterface) && $reflection->implementsInterface($translatableFactoryInterface)) {
+        if (in_array(TranslatableFactoryInterface::class, class_implements($factoryClass))) {
             $decoratedDefinition = new Definition(Factory::class);
             $decoratedDefinition->setArguments([$modelClass]);