+++ src/Sylius/Bundle/ResourceBundle/GedmoHandler/TranslationSlugHandler.php
@@ -9,7 +9,7 @@
  * file that was distributed with this source code.
  */
 
-namespace Sylius\Bundle\TranslationBundle\GedmoHandler;
+namespace Sylius\Bundle\ResourceBundle\GedmoHandler;
 
 use Doctrine\Common\Persistence\Mapping\ClassMetadata;
 use Doctrine\Common\Persistence\ObjectManager;